backend/requirements.txt
fastapi[standard]
uvicorn
PyYAML~=6.0
pillow~=11.1
hypercorn~=0.17
prisma~=0.15
matplotlib~=3.9
pyjwt~=2.10
